refactor: extract StatementZipPathResolver from FileNameGenerator
Moves the ZIP path resolution and deduplication logic into a dedicated
StatementZipPathResolver class under Logic/Statement/Export/, which is a
more accurate namespace since the logic operates on Statement entities.
FileNameGenerator is now focused solely on generating file names.
